+++ src/main/java/com/bidnbuy/server/dto/WishlistDto.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class WishlistDto {
-    private Long auctionId;
+    private Integer auctionId;
     private Integer wishCount ;
     private boolean isLiked;
 }
